refactor(@dpc-sdp/ripple-tide-api): :truck: rename modules collection
